[refactor] Move Message Class into payload package
1. controller 패키지에 있는 message 클래스를 payload package로 이동
